feat(capabilities): generate the CLI from capability metadata
The last step of the metadata/handler split. Task capability metadata and its
zod schemas now live in @octomux/capabilities; server/registry/capabilities/task.ts
imports them and pairs each with its existing handler, unchanged.

cli/src/index.ts builds task list/get/create/start/move/delete from that
metadata, so flags derive from the schema and cannot drift. Deletes the five
hand-written commands they replace. close-task stays hand-written — task.close
has no HTTP route to proxy to.

server/orchestrator/command-schemas.ts moves into the package (pure zod, no
server deps) rather than being duplicated.

Accepted breaking changes: positional args become flags (get-task <id> ->
get-task --id <id>), and create-task --fork-from, list-tasks --status and the
ls/info aliases are gone.

Verified: cd cli && tsc --noEmit is 0 errors, cli imports nothing from server/,
and bun run cli:build succeeds — the production build path that tsc -b and
vitest both miss.
